backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "consultantai-backend"
version = "0.1.0"
description = "ConsultantAI Backend API"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.23.2",
    "pydantic==2.4.2",
    "langchain>=0.1.0",
    "langchain-core>=0.1.38,<0.2.0",
    "langchain-experimental==0.0.42",
    "langgraph==0.0.33",
    "langchain-anthropic==0.1.1",
    "langsmith>=0.1.0,<0.2.0",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "bcrypt==4.0.1",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    "asyncpg==0.28.0",
    "pytest==7.4.3",
    "httpx==0.25.1",
    "pydantic-settings==2.0.3",
    "aiofiles==23.2.1",
]

[tool.setuptools.packages.find]
exclude = ["*"] 